Profile form working / Added PHP
The "edit student form" is now working on the profile.php page. I tested
it multiple times, and as of now it is working. I am unsure of how you
want to ID the student (user) using the page. I suspect this is where
the session IDs come into play. Of course, this info is needed to
determine which student object's attributes are updated in the DB. For
now, I am just using the generic first and last names "John Smith" to ID
the user for testing purposes. Obviously, this needs to be changed in
the future.